working on view matrix

+++ TraZer/Utilities/tzBBox.cpp
@@ -10,9 +10,9 @@ tzBBox::tzBBox (void)
 
 
 // ================================================================================
-tzBBox::tzBBox (	const double _x0, const double _x1,			
-				const double _y0, const double _y1, 
-				const double _z0, const double _z1)
+tzBBox::tzBBox (	const float _x0, const float _x1,			
+				const float _y0, const float _y1, 
+				const float _z0, const float _z1)
 	: x0(_x0), x1(_x1), y0(_y0), y1(_y1), z0(_z0), z1(_z1)
 {}
 
@@ -56,14 +56,14 @@ tzBBox::~tzBBox(void)
 // ================================================================================
 bool tzBBox::hit(const tzRay& ray) const 
 {
-	double ox = ray.o.x; double oy = ray.o.y; double oz = ray.o.z;
-	double dx = ray.d.x; double dy = ray.d.y; double dz = ray.d.z;
+	float ox = ray.o.x; float oy = ray.o.y; float oz = ray.o.z;
+	float dx = ray.d.x; float dy = ray.d.y; float dz = ray.d.z;
 	
-	double tx_min, ty_min, tz_min;
-	double tx_max, ty_max, tz_max; 
+	float tx_min, ty_min, tz_min;
+	float tx_max, ty_max, tz_max; 
 
-	double a = 1.0 / dx;
-	if (a >= 0) {
+	float a = 1.0f / dx;
+	if (a >= 0.0f) {
 		tx_min = (x0 - ox) * a;
 		tx_max = (x1 - ox) * a;
 	}
@@ -72,8 +72,8 @@ bool tzBBox::hit(const tzRay& ray) const
 		tx_max = (x0 - ox) * a;
 	}
 	
-	double b = 1.0 / dy;
-	if (b >= 0) {
+	float b = 1.0f / dy;
+	if (b >= 0.0f) {
 		ty_min = (y0 - oy) * b;
 		ty_max = (y1 - oy) * b;
 	}
@@ -82,7 +82,7 @@ bool tzBBox::hit(const tzRay& ray) const
 		ty_max = (y0 - oy) * b;
 	}
 	
-	double c = 1.0 / dz;
+	float c = 1.0f / dz;
 	if (c >= 0) {
 		tz_min = (z0 - oz) * c;
 		tz_max = (z1 - oz) * c;
@@ -92,7 +92,7 @@ bool tzBBox::hit(const tzRay& ray) const
 		tz_max = (z0 - oz) * c;
 	}
 	
-	double t0, t1;
+	float t0, t1;
 	
 	// find largest entering t value
